feat: 메뉴 컨텍스트 분리
