+++ kitchenpos-menu/src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,0 +1,27 @@
+package kitchenpos.menu.application;
+
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroup create(final MenuGroup request) {
+        return menuGroupRepository.save(request);
+    }
+
+    @Transactional(readOnly = true)
+    public List<MenuGroup> list() {
+        return menuGroupRepository.findAll();
+    }
+}
