+++ kitchenpos-menu/src/main/java/kitchenpos/menu/dto/MenuGroupCreateRequest.java
@@ -1,0 +1,17 @@
+package kitchenpos.menu.dto;
+
+public class MenuGroupCreateRequest {
+
+    private String name;
+
+    public MenuGroupCreateRequest(String name) {
+        this.name = name;
+    }
+
+    protected MenuGroupCreateRequest() {
+    }
+
+    public String getName() {
+        return name;
+    }
+}
