+++ kitchenpos-menu/src/main/java/kitchenpos/menu/ui/MenuGroupRestController.java
@@ -1,0 +1,39 @@
+package kitchenpos.menu.ui;
+
+import kitchenpos.menu.application.MenuGroupService;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.dto.MenuGroupCreateRequest;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class MenuGroupRestController {
+    private final MenuGroupService menuGroupService;
+
+    public MenuGroupRestController(final MenuGroupService menuGroupService) {
+        this.menuGroupService = menuGroupService;
+    }
+
+    @PostMapping("/api/menu-groups")
+    public ResponseEntity<MenuGroup> create(@RequestBody MenuGroup request) {
+        final MenuGroup created = menuGroupService.create(request);
+        final URI uri = URI.create("/api/menu-groups/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @GetMapping("/api/menu-groups")
+    public ResponseEntity<List<MenuGroup>> list() {
+        return ResponseEntity.ok()
+                .body(menuGroupService.list())
+                ;
+    }
+}
