+++ kitchenpos-product/src/main/java/kitchenpos/product/domain/Name.java
@@ -1,0 +1,10 @@
+package kitchenpos.product.domain;
+
+public interface Name {
+
+    default void validLength(String name) {
+        if (name.isEmpty()) {
+            throw new IllegalArgumentException("이름은 한 글자 이상이어야 합니다");
+        }
+    }
+}
